+++ src/components/PageHeader/index.vue
@@ -1,0 +1,19 @@
+<template>
+  <div class="page-header-container">
+    <slot name="content"></slot>
+  </div>
+</template>
+
+<script>
+export default {
+  data() {
+    return {};
+  },
+
+  computed: {},
+
+  mounted() {},
+
+  methods: {},
+};
+</script>