+++ CHANGELOG.md
@@ -1,1 +1,12 @@
-HELLO WORLD
+# v1.0.2-alpha.3(2023-03-01)
+
+### Feature
+- add workspace page
+- add ayalysis page
+- add footer component
+- add i18n components and config
+- update headers and breadCrumb
+
+### Bug Fixes
+- fix style with rem
+- fix menu style